Allow for re-doing validation and waiting for submitted jobs to complete
A new option '--validate-only' for the compute command allows one to re-run the validation phase. In our case, we use the validation phase to gather benchmark results. The option is useful, for example, if the script used to parse computation output changes and we need to reparse the latter.
Then, the second new option '--wait-after-submitting' can make gcvb wait after submitting a job for the latter to complete before submitting the next one. This allows us to prevent the cases when job.sh is concurrently overwritten by gcvb when making multiple successive calls to gcvb using the compute command.